gpt-sonarqube-fn/requirements.txt
azure-functions
openai==0.28

